refactor: rename my_error to illegal_argument_error
